requirement.txt
streamlit==1.35.0
pandas==2.2.2
plotly==5.21.0
mysql-connector-python==8.4.0